requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
psycopg2-binary==2.9.6
boto3==1.26.144
openai
requests==2.31.0
python-jose==3.3.0
python-multipart==0.0.5
Pillow==9.5.0
jwt